requirements.txt
runpod
boto3
pydantic
requests
torch
torchvision
diffusers
transformers
accelerate
sentencepiece
imageio[ffmpeg]
Pillow
